Create control.cgi
This is a work in progress.  The idea is to have a basic mobile app that can be used to monitor and control the Menuboard.  Currently the only working functionality is turning the tv off and on and viewing the screenshot.  The live view depends on the screenshot script to be running.